+++ posthog-core/test/posthog.capture.spec.ts
@@ -43,5 +43,33 @@ describe('PostHog Core', () => {
         sent_at: '2022-01-01T00:00:00.000Z',
       })
     })
+
+    it('should allow overridding the timestamp', async () => {
+      jest.setSystemTime(new Date('2022-01-01'))
+
+      posthog.capture('custom-event', { foo: 'bar' }, { timestamp: new Date('2021-01-02') })
+      const body = parseBody(mocks.fetch.mock.calls[0])
+
+      expect(body).toEqual({
+        api_key: 'TEST_API_KEY',
+        batch: [
+          {
+            event: 'custom-event',
+            distinct_id: posthog.getDistinctId(),
+            library: 'posthog-core-tests',
+            library_version: '2.0.0-alpha',
+            properties: {
+              $lib: 'posthog-core-tests',
+              $lib_version: '2.0.0-alpha',
+              $session_id: expect.any(String),
+              foo: 'bar',
+            },
+            timestamp: '2021-01-02T00:00:00.000Z',
+            type: 'capture',
+          },
+        ],
+        sent_at: '2022-01-01T00:00:00.000Z',
+      })
+    })
   })
 })